split variable query editor
