--- src/datasources/data-frame/components/DataFrameVariableQueryEditor.tsx
@@ -1,33 +1,0 @@
-import React, { useState } from 'react';
-import { AsyncSelect } from '@grafana/ui';
-import { InlineField } from 'core/components/InlineField';
-import { toOption } from '@grafana/data';
-import { isValidId } from '../utils';
-import { FloatingError, parseErrorMessage } from '../../../core/errors';
-import { DataFrameQueryEditorCommon, Props } from './DataFrameQueryEditorCommon';
-
-export function DataFrameVariableQueryEditor(props: Props) {
-  const [errorMsg, setErrorMsg] = useState<string | undefined>('');
-  const handleError = (error: Error) => setErrorMsg(parseErrorMessage(error));
-  const common = new DataFrameQueryEditorCommon(props, handleError);
-
-  return (
-    <div style={{ position: 'relative' }}>
-      <InlineField label="Id">
-        <AsyncSelect
-          allowCreateWhileLoading
-          allowCustomValue
-          cacheOptions={false}
-          defaultOptions
-          isValidNewOption={isValidId}
-          loadOptions={common.handleLoadOptions}
-          onChange={common.handleIdChange}
-          placeholder="Search by name or enter id"
-          width={30}
-          value={common.query.tableId ? toOption(common.query.tableId) : null}
-        />
-      </InlineField>
-      <FloatingError message={errorMsg}/>
-    </div>
-  );
-}
