+++ src/datasources/data-frame/components/v2/DataFrameVariableQueryEditorV2.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+export const DataFrameVariableQueryEditorV2 = () => {
+    // TODO AB#3259790: Add `Data table properties` Query Builder
+
+    return (
+        <div>
+            <p>DataFrame Variable Query Editor V2 is under development.</p>
+        </div>
+    );
+}
