+++ src/datasources/data-frame/module.ts
@@ -1,10 +1,10 @@
 import { DataSourcePlugin } from '@grafana/data';
 import { DataFrameDataSource } from './DataFrameDataSource';
 import { DataFrameQueryEditorWrapper } from './components/DataFrameQueryEditorWrapper';
-import { DataFrameVariableQueryEditor } from './components/DataFrameVariableQueryEditor';
+import { DataFrameVariableQueryEditorWrapper } from './components/DataFrameVariableQueryEditorWrapper';
 import { DataFrameConfigEditor } from './components/DataFrameConfigEditor';
 
 export const plugin = new DataSourcePlugin(DataFrameDataSource)
   .setConfigEditor(DataFrameConfigEditor)
   .setQueryEditor(DataFrameQueryEditorWrapper)
-  .setVariableQueryEditor(DataFrameVariableQueryEditor);
+  .setVariableQueryEditor(DataFrameVariableQueryEditorWrapper);
